changed the hero layout
